+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Expose Admin API with the currently logged in accounts [#355](https://github.com/rokwire/core-building-block/issues/355)
 - Add app config endpoints [#261](https://github.com/rokwire/core-building-block/issues/261)
 
+### Security
+- Fix security vulnerability for roles and groups admin APIs [#414](https://github.com/rokwire/core-building-block/issues/414)
+
 ### Fixed
 - Fix issues with account linking [#408](https://github.com/rokwire/core-building-block/issues/408)
 - Fix creating application group admin API [#397](https://github.com/rokwire/core-building-block/issues/397)